Add a CMake build
We don't want to waste time hand hacking compiler flags.  The better way
is to let CMake (or a similar tool) do the hard work.  As a bonus, we
can make sure NetCDF and FFTW are installed (and located) correctly.
The primary output is just the cloudgen executable.  If FFTW or NetCDF
is necessary, those get built and installed too.

FFTW ships with pkg-config, so I located a CMake find script.  I had to
patch it to use the fancy transient dependency resolution.  The patch is
included here, but I submitted a pull request.  The patch may break in
the future.

NetCDF now ships with a CMake!  The down side is they do a poor job
if you want to include it as an External Project.  To get it to work as
an External Project, we have to patch the CMakeLists.txt to not use
`CMAKE_SOURCE_DIR`.  This should be `CMAKE_CURRENT_SOURCE_DIR`.