Add initial Bison/Flex parsing
The hand rolled parser occasionally runs into a segmentation fault.
This appears to be due to intermittent problems with allocating the
storage.  Instead of hand rolling the parser, we can just use Bison and
Flex to do the work.  This adds the base parser but needs finalization
and a test.