Install all dependencies for linting
We need to make sure all the dependencies are installed into the virtual
environment to satisfy mypy.  That means we need to pull the details
from the setup.cfg, but we can also place the nox dependency where it
belongs in the setup.cfg.  And, we can stop telling mypy to ignore
missing imports.