Force the full path for the installed utilities
Per the documentation for `RPATH` from CMake [1], the only way to force
the installation to ignore the dynamic library search paths is to
explicitly set `RPATH` on installation.  The bulk details are we must
follow the standard CMake idioms to get the behavior we expect [2].  The
advantage of this structure is the end user does not have to care about
library paths or installation paths or anything like that.

[1]: https://gitlab.kitware.com/cmake/community/-/wikis/doc/cmake/RPATH-handling#always-full-rpath
[2]: https://stackoverflow.com/a/70163027/4249913