Force the order of random number generation
We cannot rely on the order of function evaluation to provide consistent
results.  In fact, the order in which a statement is evaluated is not
defined [1].  This prevents us from having reliable regression testing
when explicitly setting the seed which we need to ensure we are not
breaking the code when making changes.

[1]: https://en.cppreference.com/w/c/language/eval_order