Prevent nccmp from installing
We only need nccmp for testing purposes, and we do not want to install
it.  Contrary to what [FetchContent_Declare] says, the install step is
not respected, and the solutions on StackOverflow to use
EXCLUDE_FROM_ALL with add_subdirectory does not build the executable and
setting INSTALL_COMMAND to "" does not work because FetchContent_Declare
ignores it.  Instead, we just fix the version of nccmp and patch out the
install commands.

[FecthContent_Declare]: https://cmake.org/cmake/help/v3.11/module/FetchContent.html#declaring-content-details