Check for a valid version of FFTW before building
We already know we need version 3.3.4 or later of FFTW.  Therefore, we
need to check the version before continuing the tests.  Otherwise, the
build will fail.  This checks the output of fftw-wisdom-to-conf for the
version and then checks if we are using a conda environment in which it
can install FFTW.

This does require bumping the nox version to support the warn method of
the session.