Stop using a git tag to get nccmp
Using git to pull a particular tag or branch causes git to try to update
the repo every invocation of cmake.  This means the patch is being
applied _every_ time cmake is run which is bad.  Instead, we can just
pull the tar ball from a URL which the documentation even states is the
preferred way to use FetchContent.